+++ packages/client-config/src/client-config-contributor/client_config_contributor_v1.ts
@@ -93,8 +93,19 @@ export class AuthClientConfigContributor implements ClientConfigContributor {
     );
 
     if (authOutput.payload.mfaConfiguration) {
-      authClientConfig.auth.mfa_configuration = authOutput.payload
-        .mfaConfiguration as 'NONE' | 'OPTIONAL' | 'REQUIRED';
+      switch (authOutput.payload.mfaConfiguration) {
+        case 'OFF': {
+          authClientConfig.auth.mfa_configuration = 'NONE';
+          break;
+        }
+        case 'OPTIONAL': {
+          authClientConfig.auth.mfa_configuration = 'OPTIONAL';
+          break;
+        }
+        case 'ON': {
+          authClientConfig.auth.mfa_configuration = 'REQUIRED';
+        }
+      }
     }
 
     if (
@@ -135,10 +146,14 @@ export class AuthClientConfigContributor implements ClientConfigContributor {
       authOutput.payload.oauthRedirectSignIn &&
       authOutput.payload.oauthRedirectSignOut
     ) {
+      let socialProviders = authOutput.payload.socialProviders
+        ? JSON.parse(authOutput.payload.socialProviders)
+        : [];
+      if (Array.isArray(socialProviders)) {
+        socialProviders = socialProviders.filter(this.isValidIdentityProvider);
+      }
       authClientConfig.auth.oauth = {
-        identity_providers: authOutput.payload.socialProviders
-          ? JSON.parse(authOutput.payload.socialProviders)
-          : [],
+        identity_providers: socialProviders,
         redirect_sign_in_uri: authOutput.payload.oauthRedirectSignIn.split(','),
         redirect_sign_out_uri:
           authOutput.payload.oauthRedirectSignOut.split(','),
@@ -156,6 +171,16 @@ export class AuthClientConfigContributor implements ClientConfigContributor {
     }
 
     return authClientConfig;
+  };
+
+  // Define a type guard function to check if a value is a valid IdentityProvider
+  isValidIdentityProvider = (identityProvider: string): boolean => {
+    return [
+      'GOOGLE',
+      'FACEBOOK',
+      'LOGIN_WITH_AMAZON',
+      'SIGN_IN_WITH_APPLE',
+    ].includes(identityProvider);
   };
 }