+++ packages/integration-tests/src/test-project-setup/create_empty_amplify_project.ts
@@ -2,12 +2,13 @@ import fs from 'fs/promises';
 import path from 'path';
 import { shortUuid } from '../short_uuid.js';
 import { setupDirAsEsmModule } from './setup_dir_as_esm_module.js';
+import { setupDeployedBackendClient } from './setup_deployed_backend_client.js';
 
 const TEST_PROJECT_PREFIX = 'test-project';
 
 /**
  * Creates an empty Amplify project directory within the specified parent
- * The project contains an empty `amplify` directory and a package.json file with a name
+ * The project contains empty `amplify` and `.amplify` directories and a package.json file
  */
 export const createEmptyAmplifyProject = async (
   projectDirName: string,
@@ -22,7 +23,7 @@ export const createEmptyAmplifyProject = async (
   const projectName = `${TEST_PROJECT_PREFIX}-${projectDirName}-${shortUuid()}`;
   await fs.writeFile(
     path.join(projectRoot, 'package.json'),
-    JSON.stringify({ name: projectName }, null, 2)
+    JSON.stringify({ name: projectName, type: 'module' }, null, 2)
   );
 
   const projectAmplifyDir = path.join(projectRoot, 'amplify');
@@ -32,6 +33,8 @@ export const createEmptyAmplifyProject = async (
   await fs.mkdir(projectDotAmplifyDir);
 
   await setupDirAsEsmModule(projectAmplifyDir);
+
+  await setupDeployedBackendClient(projectRoot);
 
   return { projectName, projectRoot, projectAmplifyDir, projectDotAmplifyDir };
 };