+++ packages/backend-function/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @aws-amplify/backend-function
 
+## 1.2.0
+
+### Minor Changes
+
+- 63a881f: Add the ability to add function environment variables after defining the function
+
+### Patch Changes
+
+- 90ac407: Minify lambda function code and add source maps
+
+## 1.1.0
+
+### Minor Changes
+
+- 0200d11: Bundle locally available AWS SDK packages by default
+
+### Patch Changes
+
+- 0200d11: Bump baseline CDK version to 2.132.0 to support AWS SDK bundling
+- Updated dependencies [0200d11]
+  - @aws-amplify/backend-output-storage@1.0.2
+  - @aws-amplify/plugin-types@1.0.1
+
 ## 1.0.3
 
 ### Patch Changes